[Control][NMPC] in config/ folder, refactor code, separate physical parameters into a separate yaml file.
