+++ aerial_robot_control/scripts/nmpc/tilt_qd/arxiv_tilt_qd_servo_drag_w_dist.py
@@ -20,14 +20,18 @@ from nmpc_base import NMPCBase, XrUrConverterBase
 
 # read parameters from yaml
 rospack = rospkg.RosPack()
-param_path = os.path.join(rospack.get_path("beetle"), "config", "BeetleNMPCFullITermDrag.yaml")
-with open(param_path, "r") as f:
-    param_dict = yaml.load(f, Loader=yaml.FullLoader)
 
-nmpc_params = param_dict["controller"]["nmpc"]
+physical_param_path = os.path.join(rospack.get_path("beetle"), "config", "PhysParamBeetleArt.yaml")
+with open(physical_param_path, "r") as f:
+    physical_param_dict = yaml.load(f, Loader=yaml.FullLoader)
+physical_params = physical_param_dict["physical"]
+
+nmpc_param_path = os.path.join(rospack.get_path("beetle"), "config", "BeetleNMPCFullITermDrag.yaml")
+with open(nmpc_param_path, "r") as f:
+    nmpc_param_dict = yaml.load(f, Loader=yaml.FullLoader)
+nmpc_params = nmpc_param_dict["controller"]["nmpc"]
 nmpc_params["N_node"] = int(nmpc_params["T_pred"] / nmpc_params["T_integ"])
 
-physical_params = param_dict["controller"]["physical"]
 mass = physical_params["mass"]
 gravity = physical_params["gravity"]
 Ixx = physical_params["inertia_diag"][0]
@@ -115,13 +119,13 @@ class NMPCTiltQdServoDragDist(NMPCBase):
 
         # transformation matrix
         row_1 = ca.horzcat(
-            ca.SX(1 - 2 * qy ** 2 - 2 * qz ** 2), ca.SX(2 * qx * qy - 2 * qw * qz), ca.SX(2 * qx * qz + 2 * qw * qy)
+            ca.SX(1 - 2 * qy**2 - 2 * qz**2), ca.SX(2 * qx * qy - 2 * qw * qz), ca.SX(2 * qx * qz + 2 * qw * qy)
         )
         row_2 = ca.horzcat(
-            ca.SX(2 * qx * qy + 2 * qw * qz), ca.SX(1 - 2 * qx ** 2 - 2 * qz ** 2), ca.SX(2 * qy * qz - 2 * qw * qx)
+            ca.SX(2 * qx * qy + 2 * qw * qz), ca.SX(1 - 2 * qx**2 - 2 * qz**2), ca.SX(2 * qy * qz - 2 * qw * qx)
         )
         row_3 = ca.horzcat(
-            ca.SX(2 * qx * qz - 2 * qw * qy), ca.SX(2 * qy * qz + 2 * qw * qx), ca.SX(1 - 2 * qx ** 2 - 2 * qy ** 2)
+            ca.SX(2 * qx * qz - 2 * qw * qy), ca.SX(2 * qy * qz + 2 * qw * qx), ca.SX(1 - 2 * qx**2 - 2 * qy**2)
         )
         rot_ib = ca.vertcat(row_1, row_2, row_3)
 
@@ -160,10 +164,10 @@ class NMPCTiltQdServoDragDist(NMPCBase):
         dr_a2 = dr2 * a2
         dr_a3 = dr3 * a3
         dr_a4 = dr4 * a4
-        fd1 = (c4 * dr_a1 ** 4 + c3 * dr_a1 ** 3 + c2 * dr_a1 ** 2 + c1 * dr_a1 + c0) * ft1
-        fd2 = (c4 * dr_a2 ** 4 + c3 * dr_a2 ** 3 + c2 * dr_a2 ** 2 + c1 * dr_a2 + c0) * ft2
-        fd3 = (c4 * dr_a3 ** 4 + c3 * dr_a3 ** 3 + c2 * dr_a3 ** 2 + c1 * dr_a3 + c0) * ft3
-        fd4 = (c4 * dr_a4 ** 4 + c3 * dr_a4 ** 3 + c2 * dr_a4 ** 2 + c1 * dr_a4 + c0) * ft4
+        fd1 = (c4 * dr_a1**4 + c3 * dr_a1**3 + c2 * dr_a1**2 + c1 * dr_a1 + c0) * ft1
+        fd2 = (c4 * dr_a2**4 + c3 * dr_a2**3 + c2 * dr_a2**2 + c1 * dr_a2 + c0) * ft2
+        fd3 = (c4 * dr_a3**4 + c3 * dr_a3**3 + c2 * dr_a3**2 + c1 * dr_a3 + c0) * ft3
+        fd4 = (c4 * dr_a4**4 + c3 * dr_a4**3 + c2 * dr_a4**2 + c1 * dr_a4 + c0) * ft4
 
         ft_r1 = ca.vertcat(0, 0, ft1 - fd1)
         ft_r2 = ca.vertcat(0, 0, ft2 - fd2)
@@ -176,20 +180,20 @@ class NMPCTiltQdServoDragDist(NMPCBase):
         tau_r4 = ca.vertcat(0, 0, -dr4 * (ft4 - fd4) * kq_d_kt)
 
         f_u_b = (
-                ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, ft_r1))
-                + ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, ft_r2))
-                + ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, ft_r3))
-                + ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, ft_r4))
+            ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, ft_r1))
+            + ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, ft_r2))
+            + ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, ft_r3))
+            + ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, ft_r4))
         )
         tau_u_b = (
-                ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, tau_r1))
-                + ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, tau_r2))
-                + ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, tau_r3))
-                + ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, tau_r4))
-                + ca.cross(np.array(p1_b), ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, ft_r1)))
-                + ca.cross(np.array(p2_b), ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, ft_r2)))
-                + ca.cross(np.array(p3_b), ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, ft_r3)))
-                + ca.cross(np.array(p4_b), ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, ft_r4)))
+            ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, tau_r1))
+            + ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, tau_r2))
+            + ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, tau_r3))
+            + ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, tau_r4))
+            + ca.cross(np.array(p1_b), ca.mtimes(rot_be1, ca.mtimes(rot_e1r1, ft_r1)))
+            + ca.cross(np.array(p2_b), ca.mtimes(rot_be2, ca.mtimes(rot_e2r2, ft_r2)))
+            + ca.cross(np.array(p3_b), ca.mtimes(rot_be3, ca.mtimes(rot_e3r3, ft_r3)))
+            + ca.cross(np.array(p4_b), ca.mtimes(rot_be4, ca.mtimes(rot_e4r4, ft_r4)))
         )
 
         # dynamic model
@@ -239,8 +243,9 @@ class NMPCTiltQdServoDragDist(NMPCBase):
 
         # get file path for acados
         rospack = rospkg.RosPack()
-        folder_path = os.path.join(rospack.get_path("aerial_robot_control"), "include", "aerial_robot_control", "nmpc",
-                                   ocp_model.name)
+        folder_path = os.path.join(
+            rospack.get_path("aerial_robot_control"), "include", "aerial_robot_control", "nmpc", ocp_model.name
+        )
         self._mkdir(folder_path)
         os.chdir(folder_path)
